Add user-helpers namespace
This replaces the functions normally defined directly in 'user'
namespace by templates such as
https://github.com/stuartsierra/reloaded

Instead of defining these functions in the 'user' namespace, they can
be :require :refer'd from the 'user' namespace. This ensures they will
not be reloaded and will still be available after an exception during
refresh/reset.